+++ lib/core/models/editor_configs/helper_lines_configs.dart
@@ -5,39 +5,6 @@ export '../styles/helper_line_style.dart';
 /// The `HelperLineConfigs` class defines the settings for displaying helper
 /// lines in the image editor.
 /// Helper lines are used to guide users in positioning and rotating layers.
-///
-/// Usage:
-///
-/// ```dart
-/// HelperLineConfigs helperLines = HelperLineConfigs(
-///   showVerticalLine: true,
-///   showHorizontalLine: true,
-///   showRotateLine: true,
-/// );
-/// ```
-///
-/// Properties:
-///
-/// - `showVerticalLine`: Specifies whether to show the vertical helper line.
-///
-/// - `showHorizontalLine`: Specifies whether to show the horizontal helper
-///   line.
-///
-/// - `showRotateLine`: Specifies whether to show the rotate helper line.
-///
-/// Example Usage:
-///
-/// ```dart
-/// HelperLineConfigs helperLines = HelperLineConfigs(
-///   showVerticalLine: true,
-///   showHorizontalLine: true,
-///   showRotateLine: true,
-/// );
-///
-/// bool showVerticalLine = helperLines.showVerticalLine;
-/// bool showHorizontalLine = helperLines.showHorizontalLine;
-/// // Access other helper lines settings...
-/// ```
 class HelperLineConfigs {
   /// Creates an instance of the `HelperLines` class with the specified
   /// settings.
@@ -46,6 +13,7 @@ class HelperLineConfigs {
     this.showHorizontalLine = true,
     this.showRotateLine = true,
     this.showLayerAlignLine = true,
+    this.isDisabledAtZoom = false,
     this.releaseThreshold = 10.0,
     this.style = const HelperLineStyle(),
   });
@@ -61,6 +29,15 @@ class HelperLineConfigs {
 
   /// Specifies whether to show the layer align helper line.
   final bool showLayerAlignLine;
+
+  /// Determines whether the helper lines are disabled when the editor is
+  /// zoomed in.
+  ///
+  /// If set to `true`, helper lines will not be displayed when the zoom level
+  /// is increased.
+  /// If set to `false`, helper lines will remain visible regardless of the
+  /// zoom level.
+  final bool isDisabledAtZoom;
 
   /// Style configuration for helper lines.
   final HelperLineStyle style;
@@ -80,6 +57,7 @@ class HelperLineConfigs {
     bool? showHorizontalLine,
     bool? showRotateLine,
     bool? showLayerAlignLine,
+    bool? isDisabledAtZoom,
     double? releaseThreshold,
     HelperLineStyle? style,
   }) {
@@ -88,6 +66,7 @@ class HelperLineConfigs {
       showHorizontalLine: showHorizontalLine ?? this.showHorizontalLine,
       showRotateLine: showRotateLine ?? this.showRotateLine,
       showLayerAlignLine: showLayerAlignLine ?? this.showLayerAlignLine,
+      isDisabledAtZoom: isDisabledAtZoom ?? this.isDisabledAtZoom,
       releaseThreshold: releaseThreshold ?? this.releaseThreshold,
       style: style ?? this.style,
     );