+++ lib/core/models/editor_callbacks/main_editor/main_editor_callbacks.dart
@@ -2,6 +2,7 @@
 import 'package:flutter/widgets.dart';
 
 import '/features/main_editor/main_editor.dart';
+import '/features/main_editor/services/state_manager.dart';
 import '/shared/services/import_export/import_state_history.dart';
 import '../../../enums/sub_editors_name.dart';
 import '../../layers/layer.dart';
@@ -40,6 +41,7 @@ class MainEditorCallbacks extends StandaloneEditorCallbacks {
     this.onImportHistoryStart,
     this.onImportHistoryEnd,
     this.onHoverRemoveAreaChange,
+    this.onStateHistoryChange,
     super.onInit,
     super.onAfterViewInit,
     super.onUpdateUI,
@@ -99,6 +101,11 @@ class MainEditorCallbacks extends StandaloneEditorCallbacks {
   /// This can be `null` if no action is required at the start of the close
   /// process.
   final Function(SubEditor editor)? onStartCloseSubEditor;
+
+  /// Callback that is triggered whenever the state history of the editor
+  /// changes.
+  final Function(StateManager stateHistory, ProImageEditorState editor)?
+      onStateHistoryChange;
 
   /// A callback function that is triggered when the user `tap` on the body.
   final Function()? onTap;
@@ -411,6 +418,8 @@ class MainEditorCallbacks extends StandaloneEditorCallbacks {
     Function(ProImageEditorState state, ImportStateHistory import)?
         onImportHistoryEnd,
     Function(bool isPointerInside)? onHoverRemoveAreaChange,
+    Function(StateManager stateHistory, ProImageEditorState editor)?
+        onStateHistoryChange,
   }) {
     return MainEditorCallbacks(
       onLayerTapDown: onLayerTapDown ?? this.onLayerTapDown,
@@ -451,6 +460,7 @@ class MainEditorCallbacks extends StandaloneEditorCallbacks {
       onImportHistoryEnd: onImportHistoryEnd ?? this.onImportHistoryEnd,
       onHoverRemoveAreaChange:
           onHoverRemoveAreaChange ?? this.onHoverRemoveAreaChange,
+      onStateHistoryChange: onStateHistoryChange ?? this.onStateHistoryChange,
     );
   }
 }